requirements.txt
aiohttp
archspec
astunparse
brotli
cairosvg
conda-build
conda-content-trust
conda-libmamba-solver
distro
expecttest
fastwer
fire
gin-config
gmpy2
hypothesis
ipython
lightning
loguru
lxml
matplotlib
mkl-fft
mkl-random
moviepy
names
opencv-python
optree
pip-chill
platformdirs
pysocks
python-dotenv
rich
scikit-image
scikit-learn
timm
toolz
torchaudio
torchelastic
torchinfo
triton
types-dataclasses
verovio
wand
wandb
wonderwords
